requirements.txt
python-dotenv
python-telegram-bot
selenium
sqlalchemy
xlwings